# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Supply Chain Optimization System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: shipments
    description: Shipment tracking and management
  - name: routes
    description: Route optimization
  - name: disruptions
    description: Disruption alerts and risk management
  - name: analytics
    description: Predictive analytics and dashboard summaries
  - name: warehouses
    description: Warehouse and port management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List all shipments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_transit, delayed, delivered, at_risk]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a new shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentBody"
      responses:
        "201":
          description: Created shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get a shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shipment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateShipment
      tags: [shipments]
      summary: Update shipment status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentBody"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
    delete:
      operationId: deleteShipment
      tags: [shipments]
      summary: Delete a shipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /disruptions:
    get:
      operationId: listDisruptions
      tags: [disruptions]
      summary: List disruption alerts
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of disruptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disruption"
    post:
      operationId: createDisruption
      tags: [disruptions]
      summary: Report a new disruption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisruptionBody"
      responses:
        "201":
          description: Created disruption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disruption"

  /disruptions/{id}:
    patch:
      operationId: updateDisruption
      tags: [disruptions]
      summary: Resolve or update a disruption
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDisruptionBody"
      responses:
        "200":
          description: Updated disruption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disruption"

  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List optimized routes
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteBody"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"

  /routes/{id}/optimize:
    post:
      operationId: optimizeRoute
      tags: [routes]
      summary: Get optimized alternative for a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Optimization result with alternatives
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteOptimizationResult"

  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses and ports
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"

  /analytics/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [analytics]
      summary: Get dashboard KPI summary
      responses:
        "200":
          description: Dashboard summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /analytics/risk-scores:
    get:
      operationId: getRiskScores
      tags: [analytics]
      summary: Get risk scores for active shipments
      responses:
        "200":
          description: Risk scores per shipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskScore"

  /analytics/delay-forecast:
    get:
      operationId: getDelayForecast
      tags: [analytics]
      summary: Get delay forecasts for next 7 days
      responses:
        "200":
          description: Delay forecast data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DelayForecast"

  /analytics/disruption-trends:
    get:
      operationId: getDisruptionTrends
      tags: [analytics]
      summary: Get disruption trends over time
      responses:
        "200":
          description: Disruption trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisruptionTrend"

  /analytics/cost-breakdown:
    get:
      operationId: getCostBreakdown
      tags: [analytics]
      summary: Get cost vs time optimization breakdown
      responses:
        "200":
          description: Cost breakdown by transport mode
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CostBreakdown"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Shipment:
      type: object
      properties:
        id:
          type: integer
        trackingId:
          type: string
        origin:
          type: string
        destination:
          type: string
        currentLocation:
          type: string
        status:
          type: string
          enum: [pending, in_transit, delayed, delivered, at_risk]
        transportMode:
          type: string
          enum: [air, sea, rail, road, multimodal]
        weight:
          type: number
        estimatedDelivery:
          type: string
          format: date-time
        actualDelivery:
          type: string
          format: date-time
          nullable: true
        riskScore:
          type: number
          minimum: 0
          maximum: 100
        delayProbability:
          type: number
          minimum: 0
          maximum: 1
        estimatedDelayhours:
          type: number
        carrier:
          type: string
        cost:
          type: number
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - trackingId
        - origin
        - destination
        - currentLocation
        - status
        - transportMode
        - weight
        - estimatedDelivery
        - riskScore
        - delayProbability
        - estimatedDelayhours
        - carrier
        - cost
        - createdAt
        - updatedAt

    CreateShipmentBody:
      type: object
      properties:
        trackingId:
          type: string
        origin:
          type: string
        destination:
          type: string
        currentLocation:
          type: string
        status:
          type: string
          enum: [pending, in_transit, delayed, delivered, at_risk]
        transportMode:
          type: string
          enum: [air, sea, rail, road, multimodal]
        weight:
          type: number
        estimatedDelivery:
          type: string
          format: date-time
        carrier:
          type: string
        cost:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        notes:
          type: string
      required:
        - trackingId
        - origin
        - destination
        - currentLocation
        - status
        - transportMode
        - weight
        - estimatedDelivery
        - carrier
        - cost

    UpdateShipmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_transit, delayed, delivered, at_risk]
        currentLocation:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        riskScore:
          type: number
        delayProbability:
          type: number
        estimatedDelayhours:
          type: number
        notes:
          type: string

    Disruption:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [weather, traffic, port_congestion, labor_strike, mechanical, geopolitical, customs, other]
        severity:
          type: string
          enum: [low, medium, high, critical]
        title:
          type: string
        description:
          type: string
        affectedRegion:
          type: string
        affectedRoutes:
          type: array
          items:
            type: string
        estimatedImpactHours:
          type: number
        resolved:
          type: boolean
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - severity
        - title
        - description
        - affectedRegion
        - affectedRoutes
        - estimatedImpactHours
        - resolved
        - createdAt
        - updatedAt

    CreateDisruptionBody:
      type: object
      properties:
        type:
          type: string
          enum: [weather, traffic, port_congestion, labor_strike, mechanical, geopolitical, customs, other]
        severity:
          type: string
          enum: [low, medium, high, critical]
        title:
          type: string
        description:
          type: string
        affectedRegion:
          type: string
        affectedRoutes:
          type: array
          items:
            type: string
        estimatedImpactHours:
          type: number
        latitude:
          type: number
        longitude:
          type: number
      required:
        - type
        - severity
        - title
        - description
        - affectedRegion
        - estimatedImpactHours

    UpdateDisruptionBody:
      type: object
      properties:
        resolved:
          type: boolean
        severity:
          type: string
          enum: [low, medium, high, critical]
        estimatedImpactHours:
          type: number
        description:
          type: string

    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        transportMode:
          type: string
          enum: [air, sea, rail, road, multimodal]
        distanceKm:
          type: number
        estimatedHours:
          type: number
        costPerKg:
          type: number
        carbonFootprint:
          type: number
        reliability:
          type: number
          minimum: 0
          maximum: 100
        isOptimal:
          type: boolean
        waypoints:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - origin
        - destination
        - transportMode
        - distanceKm
        - estimatedHours
        - costPerKg
        - carbonFootprint
        - reliability
        - isOptimal
        - waypoints
        - createdAt

    CreateRouteBody:
      type: object
      properties:
        name:
          type: string
        origin:
          type: string
        destination:
          type: string
        transportMode:
          type: string
          enum: [air, sea, rail, road, multimodal]
        distanceKm:
          type: number
        estimatedHours:
          type: number
        costPerKg:
          type: number
        carbonFootprint:
          type: number
        reliability:
          type: number
        waypoints:
          type: array
          items:
            type: string
      required:
        - name
        - origin
        - destination
        - transportMode
        - distanceKm
        - estimatedHours
        - costPerKg
        - carbonFootprint
        - reliability

    RouteOptimizationResult:
      type: object
      properties:
        originalRoute:
          $ref: "#/components/schemas/Route"
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/RouteAlternative"
        recommendation:
          type: string
        costSaving:
          type: number
        timeSaving:
          type: number
      required:
        - originalRoute
        - alternatives
        - recommendation
        - costSaving
        - timeSaving

    RouteAlternative:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/Route"
        reason:
          type: string
        tradeoffs:
          type: string
        score:
          type: number
      required:
        - route
        - reason
        - tradeoffs
        - score

    Warehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [warehouse, port, airport, rail_terminal, distribution_center]
        city:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        congestionLevel:
          type: string
          enum: [low, medium, high, critical]
        capacity:
          type: number
        currentOccupancy:
          type: number
        operationalStatus:
          type: string
          enum: [operational, limited, closed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - city
        - country
        - latitude
        - longitude
        - congestionLevel
        - capacity
        - currentOccupancy
        - operationalStatus
        - createdAt

    DashboardSummary:
      type: object
      properties:
        totalShipments:
          type: integer
        activeShipments:
          type: integer
        delayedShipments:
          type: integer
        atRiskShipments:
          type: integer
        deliveredToday:
          type: integer
        activeDisruptions:
          type: integer
        criticalDisruptions:
          type: integer
        avgRiskScore:
          type: number
        onTimeRate:
          type: number
        totalCost:
          type: number
        potentialSavings:
          type: number
      required:
        - totalShipments
        - activeShipments
        - delayedShipments
        - atRiskShipments
        - deliveredToday
        - activeDisruptions
        - criticalDisruptions
        - avgRiskScore
        - onTimeRate
        - totalCost
        - potentialSavings

    RiskScore:
      type: object
      properties:
        shipmentId:
          type: integer
        trackingId:
          type: string
        riskScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        riskFactors:
          type: array
          items:
            type: string
        delayProbability:
          type: number
        estimatedDelayhours:
          type: number
      required:
        - shipmentId
        - trackingId
        - riskScore
        - riskLevel
        - riskFactors
        - delayProbability
        - estimatedDelayhours

    DelayForecast:
      type: object
      properties:
        date:
          type: string
        forecastedDelays:
          type: integer
        predictedAvgDelayHours:
          type: number
        confidence:
          type: number
        mainCause:
          type: string
      required:
        - date
        - forecastedDelays
        - predictedAvgDelayHours
        - confidence
        - mainCause

    DisruptionTrend:
      type: object
      properties:
        date:
          type: string
        weather:
          type: integer
        traffic:
          type: integer
        port_congestion:
          type: integer
        labor_strike:
          type: integer
        mechanical:
          type: integer
        other:
          type: integer
      required:
        - date
        - weather
        - traffic
        - port_congestion
        - labor_strike
        - mechanical
        - other

    CostBreakdown:
      type: object
      properties:
        mode:
          type: string
        avgCostPerKg:
          type: number
        avgTransitHours:
          type: number
        reliability:
          type: number
        carbonFootprint:
          type: number
        shipmentCount:
          type: integer
      required:
        - mode
        - avgCostPerKg
        - avgTransitHours
        - reliability
        - carbonFootprint
        - shipmentCount