# lib/api-zod/src/generated/types/createRouteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRouteBodyTransportMode } from "./createRouteBodyTransportMode";

export interface CreateRouteBody {
  name: string;
  origin: string;
  destination: string;
  transportMode: CreateRouteBodyTransportMode;
  distanceKm: number;
  estimatedHours: number;
  costPerKg: number;
  carbonFootprint: number;
  reliability: number;
  waypoints?: string[];
}