# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalShipments: number;
  activeShipments: number;
  delayedShipments: number;
  atRiskShipments: number;
  deliveredToday: number;
  activeDisruptions: number;
  criticalDisruptions: number;
  avgRiskScore: number;
  onTimeRate: number;
  totalCost: number;
  potentialSavings: number;
}