# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScoreRiskLevel } from "./riskScoreRiskLevel";

export interface RiskScore {
  shipmentId: number;
  trackingId: string;
  riskScore: number;
  riskLevel: RiskScoreRiskLevel;
  riskFactors: string[];
  delayProbability: number;
  estimatedDelayhours: number;
}