# lib/api-zod/src/generated/types/routeOptimizationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */
import type { Route } from "./route";
import type { RouteAlternative } from "./routeAlternative";

export interface RouteOptimizationResult {
  originalRoute: Route;
  alternatives: RouteAlternative[];
  recommendation: string;
  costSaving: number;
  timeSaving: number;
}