# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */
import type { ShipmentStatus } from "./shipmentStatus";
import type { ShipmentTransportMode } from "./shipmentTransportMode";

export interface Shipment {
  id: number;
  trackingId: string;
  origin: string;
  destination: string;
  currentLocation: string;
  status: ShipmentStatus;
  transportMode: ShipmentTransportMode;
  weight: number;
  estimatedDelivery: Date;
  actualDelivery?: Date | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  riskScore: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  delayProbability: number;
  estimatedDelayhours: number;
  carrier: string;
  cost: number;
  latitude?: number | null;
  longitude?: number | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}