# lib/api-zod/src/generated/types/warehouse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Supply Chain Optimization System API
 * OpenAPI spec version: 0.1.0
 */
import type { WarehouseCongestionLevel } from "./warehouseCongestionLevel";
import type { WarehouseOperationalStatus } from "./warehouseOperationalStatus";
import type { WarehouseType } from "./warehouseType";

export interface Warehouse {
  id: number;
  name: string;
  type: WarehouseType;
  city: string;
  country: string;
  latitude: number;
  longitude: number;
  congestionLevel: WarehouseCongestionLevel;
  capacity: number;
  currentOccupancy: number;
  operationalStatus: WarehouseOperationalStatus;
  createdAt: Date;
}